+++ presentation/src/main/java/com/softpie/karabiner/ui/signup/tel/TelScreen.kt
@@ -1,22 +1,28 @@
-package com.softpie.karabiner.ui.signup.email
+package com.softpie.karabiner.ui.signup.tel
 
+import android.Manifest
 import android.app.Activity
+import android.content.Context
+import android.content.pm.PackageManager
+import android.telephony.TelephonyManager
+import android.util.Log
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
-import androidx.compose.foundation.layout.isImeVisible
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -27,6 +33,8 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
+import androidx.core.app.ActivityCompat
+import androidx.core.text.isDigitsOnly
 import androidx.core.view.WindowCompat
 import androidx.navigation.NavController
 import com.softpie.karabiner.component.button.KarabinerButton
@@ -37,7 +45,7 @@ import com.softpie.karabiner.component.theme.Label
 import com.softpie.karabiner.component.theme.Title
 import com.softpie.karabiner.local.sharedpreferences.KarabinerSharedPreferences
 import com.softpie.karabiner.ui.root.NavGroup
-import com.softpie.karabiner.utiles.isValidEmail
+import com.softpie.karabiner.utiles.isValidPhoneNumber
 import com.softpie.karabiner.utiles.rememberKeyboardIsOpen
 import com.softpie.karabiner.utiles.shortToast
 
@@ -45,7 +53,7 @@ import com.softpie.karabiner.utiles.shortToast
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
-fun SignupEmailScreen(
+fun SignupTelScreen(
     navController: NavController
 ) {
     val context = LocalContext.current
@@ -54,6 +62,26 @@ fun SignupEmailScreen(
     var value by remember { mutableStateOf("")}
     val focus = LocalFocusManager.current
     val keyboardShow by rememberKeyboardIsOpen()
+    val launcher = rememberLauncherForActivityResult(contract = ActivityResultContracts.RequestPermission()) {
+        Log.d("TAG", "Signup: $it")
+        if (ActivityCompat.checkSelfPermission(
+                context,
+                Manifest.permission.READ_PHONE_NUMBERS
+            ) != PackageManager.PERMISSION_GRANTED
+        ) {
+
+        } else {
+            val telephonyManager =
+                context.getSystemService(Context.TELEPHONY_SERVICE) as TelephonyManager
+            val number = telephonyManager.line1Number.toString().replace("+82", "0")
+            if (number.isValidPhoneNumber()) {
+                value = number
+            }
+        }
+    }
+    LaunchedEffect(true) {
+        launcher.launch(Manifest.permission.READ_PHONE_NUMBERS)
+    }
     Column(modifier = Modifier
         .fillMaxSize()
         .imePadding()
@@ -69,23 +97,25 @@ fun SignupEmailScreen(
                 .padding(horizontal = 24.dp)
         ) {
             Spacer(modifier = Modifier.height(36.dp))
-            Headline(text = "마지막으로")
+            Headline(text = "다음으로,")
             Row {
-                Title(text = "이메일", karabinerable = true)
-                Title(text = "을 입력해주세요")
+                Title(text = "전화번호", karabinerable = true)
+                Title(text = "를 입력해주세요")
             }
             Spacer(modifier = Modifier.height(8.dp))
-            Label(text = "연락이 가능한 본인의 이메일로 넣어주세요.")
+            Label(text = "연락이 가능한 본인의 번호로 입력해주세요.")
             Spacer(modifier = Modifier.height(24.dp))
             KarabinerTextField(
                 modifier = Modifier
                     .fillMaxWidth(),
-                hint = "ex) byungchoon@4rne5.dev",
+                hint = "ex) 01012345678",
                 value = value,
                 onValueChange = {
-                    value = it
+                    if (it.isDigitsOnly()) {
+                        value = it
+                    }
                 },
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email)
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
             )
         }
         Spacer(modifier = Modifier.weight(1f))
@@ -101,11 +131,11 @@ fun SignupEmailScreen(
             text = "다음으로",
             shape = if (keyboardShow) RoundedCornerShape(0.dp) else KarabinerTheme.shape.semiLarge
         ) {
-            if (value.isNotEmpty() && value.isValidEmail()) {
-                KarabinerSharedPreferences(context).myEmail = value
-                navController.navigate(NavGroup.Auth.COMPLETE.id)
+            if (value.isNotEmpty() && value.isValidPhoneNumber()) {
+                KarabinerSharedPreferences(context).myTel = value
+                navController.navigate(NavGroup.Auth.EMAIL.id)
             } else {
-                context.shortToast("이메일을 입력해주세요")
+                context.shortToast("휴대폰 번호를 입력해주세요")
             }
         }
     }
